fix: missing permissions in example
There's a problem that the action reports success even though the assets
are not actually uploaded, it's easy to miss this in logs:

```json
{"message":"Resource not accessible by integration","request_id":"<id>","documentation_url":"https://docs.github.com/rest"}
```

Ideally, the action should parse `status` and throw but my problem was
caused by missing `permission` in my workflow. Let's add it to example
to make sure ppl don't run into the same problem.